requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
Brlapi==0.6.6
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2018.1.18
cffi==1.12.3
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cssselect2==0.2.2
cupshelpers==1.0
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
Django==2.2.6
django-cors-headers==3.1.1
django-crispy-forms==1.7.2
django-qr-code==1.0.0
djangorestframework==3.10.3
et-xmlfile==1.0.1
html5lib==1.0.1
httplib2==0.9.2
idna==2.6
importlib-metadata==0.23
jdcal==1.4.1
Jinja2==2.10.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
libsass==0.19.2
louis==3.5.0
lxml==4.4.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
more-itertools==7.2.0
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
openpyxl==3.0.0
packaging==19.2
passlib==1.7.1
pexpect==4.2.1
Pillow==5.1.0
pluggy==0.13.0
protobuf==3.0.0
psycopg2-binary==2.8.3
py==1.8.0
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.2
PyPDF2==1.26.0
Pyphen==0.9.5
pyRFC3339==1.0
pytest==5.1.3
python-apt==1.6.4
python-dateutil==2.8.0
python-debian==0.1.32
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
qrcode==6.1
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
tinycss2==1.0.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
wcwidth==0.1.7
WeasyPrint==50
webencodings==0.5.1
xhtml2pdf==0.2.3
xkit==0.0.0
zipp==0.6.0
zope.interface==4.3.2
